+++ tst_pixel_bounce.py
